+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Telemetry improvements. ([@palkan][])
+
 ## 1.5.4 (2024-10-08)
 
 - Add `anycable.toml` support. ([@palkan][])